// packages/sync/src/api/generated/models/SnapshotProofChain.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Safely Sync
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SnapshotProofChain
 */
export interface SnapshotProofChain {
    /**
     * 
     * @type {Array<string>}
     * @memberof SnapshotProofChain
     */
    proofChain: Array<string>;
}

/**
 * Check if a given object implements the SnapshotProofChain interface.
 */
export function instanceOfSnapshotProofChain(value: object): value is SnapshotProofChain {
    if (!('proofChain' in value) || value['proofChain'] === undefined) return false;
    return true;
}

export function SnapshotProofChainFromJSON(json: any): SnapshotProofChain {
    return SnapshotProofChainFromJSONTyped(json, false);
}

export function SnapshotProofChainFromJSONTyped(json: any, ignoreDiscriminator: boolean): SnapshotProofChain {
    if (json == null) {
        return json;
    }
    return {
        
        'proofChain': json['proofChain'],
    };
}

export function SnapshotProofChainToJSON(json: any): SnapshotProofChain {
    return SnapshotProofChainToJSONTyped(json, false);
}

export function SnapshotProofChainToJSONTyped(value?: SnapshotProofChain | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'proofChain': value['proofChain'],
    };
}