// packages/sync/src/api/generated/models/SnapshotWithProofs.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Safely Sync
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Snapshot } from './Snapshot';
import {
    SnapshotFromJSON,
    SnapshotFromJSONTyped,
    SnapshotToJSON,
    SnapshotToJSONTyped,
} from './Snapshot';
import type { SnapshotProofChain } from './SnapshotProofChain';
import {
    SnapshotProofChainFromJSON,
    SnapshotProofChainFromJSONTyped,
    SnapshotProofChainToJSON,
    SnapshotProofChainToJSONTyped,
} from './SnapshotProofChain';

/**
 * 
 * @export
 * @interface SnapshotWithProofs
 */
export interface SnapshotWithProofs {
    /**
     * 
     * @type {Snapshot}
     * @memberof SnapshotWithProofs
     */
    snapshot: Snapshot;
    /**
     * 
     * @type {SnapshotProofChain}
     * @memberof SnapshotWithProofs
     */
    proofChain?: SnapshotProofChain;
}

/**
 * Check if a given object implements the SnapshotWithProofs interface.
 */
export function instanceOfSnapshotWithProofs(value: object): value is SnapshotWithProofs {
    if (!('snapshot' in value) || value['snapshot'] === undefined) return false;
    return true;
}

export function SnapshotWithProofsFromJSON(json: any): SnapshotWithProofs {
    return SnapshotWithProofsFromJSONTyped(json, false);
}

export function SnapshotWithProofsFromJSONTyped(json: any, ignoreDiscriminator: boolean): SnapshotWithProofs {
    if (json == null) {
        return json;
    }
    return {
        
        'snapshot': SnapshotFromJSON(json['snapshot']),
        'proofChain': json['proofChain'] == null ? undefined : SnapshotProofChainFromJSON(json['proofChain']),
    };
}

export function SnapshotWithProofsToJSON(json: any): SnapshotWithProofs {
    return SnapshotWithProofsToJSONTyped(json, false);
}

export function SnapshotWithProofsToJSONTyped(value?: SnapshotWithProofs | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'snapshot': SnapshotToJSON(value['snapshot']),
        'proofChain': SnapshotProofChainToJSON(value['proofChain']),
    };
}