// packages/sync/src/api/generated/apis/SnapshotsApi.ts
export interface GetActualSnapshotRequest {
    withProofChainTo?: string;
}

export interface GetSnapshotProofChainRequest {
    snapshotProof: string;
}

export interface SaveSnapshotRequest {
    snapshot: Snapshot;
}